refactor: 회사 공고, 지원 내역 패키지 분리
