+++ src/main/java/com/wanted/preonboarding/jobpost/repository/JobPostRepository.java
@@ -1,6 +1,6 @@
-package com.wanted.preonboarding.recruitment.repository;
+package com.wanted.preonboarding.jobpost.repository;
 
-import com.wanted.preonboarding.recruitment.entity.JobPost;
+import com.wanted.preonboarding.jobpost.entity.JobPost;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JobPostRepository extends JpaRepository<JobPost, Long>{
